chore: address CR comments
