+++ packages/bridging/src/BridgingSdk/types.ts
@@ -1,4 +1,8 @@
 import { QuoteResults, SwapAdvancedSettings, TradingAppDataInfo, TradingSdk } from '@cowprotocol/sdk-trading'
+import { SignerLike, TTLCache } from '@cowprotocol/sdk-common'
+import { TokenInfo } from '@cowprotocol/sdk-config'
+import { OrderBookApi } from '@cowprotocol/sdk-order-book'
+import { CowEnv, SupportedChainId } from '@cowprotocol/sdk-config'
 import {
   BridgeHook,
   BridgeProvider,
@@ -7,9 +11,6 @@ import {
   QuoteBridgeRequest,
   QuoteBridgeRequestWithoutAmount,
 } from '../types'
-import { SignerLike, TTLCache } from '@cowprotocol/sdk-common'
-import { BridgingSdkOptions } from './interfaces'
-import { TokenInfo } from '@cowprotocol/sdk-config'
 
 export type GetQuoteWithBridgeParams<T extends BridgeQuoteResult> = {
   /**
@@ -68,3 +69,67 @@ export interface BridgeResultContext<T extends BridgeQuoteResult = BridgeQuoteRe
 }
 
 export type BridgingSdkConfig = Required<Omit<BridgingSdkOptions, 'enableLogging' | 'cacheConfig'>>
+
+/**
+ * Cache configuration for BridgingSdk
+ */
+export interface BridgingSdkCacheConfig {
+  /**
+   * Enable caching for target networks and buy tokens
+   */
+  enabled: boolean
+  /**
+   * TTL in milliseconds for getIntermediateTokens cache
+   */
+  intermediateTokensTtl: number
+  /**
+   * TTL in milliseconds for getBuyTokens cache
+   */
+  buyTokensTtl: number
+}
+
+export interface BridgingSdkOptions {
+  /**
+   * Providers for the bridging.
+   */
+  providers: BridgeProvider<BridgeQuoteResult>[]
+
+  /**
+   * Trading SDK.
+   */
+  tradingSdk?: TradingSdk
+
+  /**
+   * Order book API.
+   */
+  orderBookApi?: OrderBookApi
+
+  /**
+   * Enable logging for the bridging SDK.
+   */
+  enableLogging?: boolean
+
+  /**
+   * Cache configuration for BridgingSdk
+   */
+  cacheConfig?: BridgingSdkCacheConfig
+}
+
+/**
+ * Parameters for the `getOrder` method.
+ */
+export interface GetOrderParams {
+  /**
+   * Id of a network where order was settled
+   */
+  chainId: SupportedChainId
+  /**
+   * The unique identifier of the order.
+   */
+  orderId: string
+
+  /**
+   * The environment of the order
+   */
+  env?: CowEnv
+}
